# Conflicts:
#	Caravela.Framework.Impl/Linking/LinkerAnalysisRegistry.cs
#	Caravela.Framework.Impl/Linking/LinkerAnnotationExtensions.cs
#	Caravela.Framework.Impl/Linking/LinkerLinkingStep.LinkingRewriter.cs
#	Caravela.Framework.Impl/Linking/LinkerLinkingStep.MethodInliningRewriter.cs
#	Caravela.Framework.Impl/Linking/LinkerLinkingStep.cs
#	Caravela.Framework.Impl/Transformations/IntroducedInterfaceImplementation.cs
#	Caravela.Framework/Aspects/AspectLinkerOptions.cs
#	Tests/Caravela.Framework.Tests.Integration/Caravela.Framework.Tests.Integration.csproj
#	Tests/Caravela.Framework.Tests.Integration/Tests/Formatting/CompileTimeIfCondition.t.cs
#	Tests/Caravela.Framework.Tests.Integration/Tests/Formatting/CompileTimeVariableDeclarations.t.cs
#	Tests/Caravela.Framework.Tests.Integration/Tests/Formatting/ForEachStatements.t.cs
#	Tests/Caravela.Framework.Tests.Integration/Tests/Formatting/RunTimeIfCondition.t.cs
#	Tests/Caravela.Framework.Tests.Integration/Tests/Formatting/RunTimeVariableDeclarations.t.cs
#	Tests/Caravela.Framework.Tests.Integration/Tests/Formatting/TemplateKeywords.t.cs

+++ Tests/Caravela.Framework.Tests.Integration/Tests/Templating/Syntax/New/NewClassCompileTime.t.cs
@@ -2,5 +2,5 @@ int Method(int a)
 {
     global::System.Console.WriteLine("string");
     global::System.Console.WriteLine("string");
-    return this.Method(a);
+    return (int)this.Method(a);
 }
