+++ Tests/Caravela.Framework.Tests.Integration/Tests/Templating/Syntax/TryCatchFinally/TryCatchFinallyCompileTime.t.cs
@@ -12,5 +12,5 @@ int Method(int a)
         global::System.Console.WriteLine(1);
     }
 
-    return this.Method(a);
+    return (int)this.Method(a);
 }
